+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 
 ### 0.16.0
  - Added support for the `all-but-first` value of the `thjs:remove` processor.
+ - Added support for `thjs:datetime` -> `datetime` HTML attribute
+ - Small performance tweaks to address 'over re-processing' of templates.
 
 ### 0.15.1
  - Fix for the `th:if` processor not causing the element to be reprocessed once